Refactored duplicate code between poly_f and poly_z into poly module
