+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod ccs;
 mod const_helpers;
 pub mod field;
 pub mod field_config;
+pub mod poly;
 pub mod poly_f;
 pub mod poly_z;
 pub mod prime_gen;
