+++ src/poly.rs
@@ -1,10 +1,10 @@
-mod multilinear_polynomial;
+pub mod errors;
+pub mod util;
 
-pub use multilinear_polynomial::{
-    evaluate_no_par, evaluate_opt, fix_last_variables, fix_last_variables_no_par, fix_variables,
-    identity_permutation, identity_permutation_mles, merge_polynomials, random_mle_list,
-    random_permutation, random_permutation_mles, random_zero_mle_list, DenseMultilinearExtension,
-};
+//TODO add exports
+pub use errors::ArithErrors;
+
+pub use util::get_batched_nv;
 
 extern crate alloc;
