+++ src/poly_z/polynomials/multilinear_polynomial.rs
@@ -3,12 +3,10 @@
 
 // Adapted for rings by Nethermind
 
+use crate::poly::{get_batched_nv, ArithErrors, RefCounter};
+pub use crate::poly_z::mle::{DenseMultilinearExtension, MultilinearExtension};
 use ark_std::{end_timer, rand::RngCore, start_timer, string::ToString, vec::*};
 use crypto_bigint::{Int, Random};
-
-use super::{util::get_batched_nv, ArithErrors, RefCounter};
-
-pub use crate::poly_z::mle::{DenseMultilinearExtension, MultilinearExtension};
 
 /// Sample a random list of multilinear polynomials.
 /// Returns
