+++ src/zinc/errors.rs
@@ -1,10 +1,8 @@
 #![allow(clippy::enum_variant_names)]
 use thiserror::Error;
 
-use crate::{
-    ccs::error::CSError, poly_f::polynomials::ArithErrors, sumcheck::SumCheckError,
-    zip::Error as ZipError,
-};
+use crate::poly::ArithErrors;
+use crate::{ccs::error::CSError, sumcheck::SumCheckError, zip::Error as ZipError};
 
 #[derive(Debug, Error)]
 pub enum ZincError<const N: usize> {
